Fixes missing test file
For an earlier PR (#16), I introduced a new test file for testing
unicode support on Windows. The project's `.gitignore` file excludes the
entire test directory, so it was missed out. Turns out that this test
file wasn't actually needed, but the CMake config still uses it. This
commit just changes this back to the original `test.cpp` file.